+++ src/cli/commands/public-cmds/test-cmd.js
@@ -10,10 +10,10 @@ import type { SpecsResultsWithComponentId } from '../../../consumer/specs-result
 const validForkLevels = R.values(TESTS_FORK_LEVEL);
 export default class Test extends Command {
   name = 'test [id]';
-  description = `test any set of components with configured tester (as defined in bit.json)\n  https://${BASE_DOCS_DOMAIN}/docs/testing-components.html`;
+  description = `test any set of components with a configured tester as defined in bit.json (by default applies only on modified components)\n  https://${BASE_DOCS_DOMAIN}/docs/testing-components.html)`;
   alias = '';
   opts = [
-    ['', 'include-unmodified', 'test all components in workspace. including unmodified'],
+    ['a', 'all', 'test all components in your workspace, including unmodified components'],
     ['v', 'verbose', 'showing npm verbose output for inspection'],
     ['', 'fork-level <forkLevel>', 'NONE / ONE / COMPONENT how many child process create for test running']
   ];
@@ -23,24 +23,24 @@ export default class Test extends Command {
   action(
     [id]: [string],
     {
-      includeUnmodified,
+      all,
       verbose,
       forkLevel
     }: {
-      includeUnmodified: ?boolean,
+      all: ?boolean,
       verbose: ?boolean,
       forkLevel: ?string
     }
   ): Promise<SpecsResultsWithComponentId> {
-    if (id && includeUnmodified) {
+    if (id && all) {
       throw new GeneralError(
-        'use --include-unmodified to test all components or use a component ID to test a specific component. run tests to all new and modified components by removing all flags and parameters'
+        'use "--all" to test all components or use a component ID to test a specific component. run tests to all new and modified components by removing all flags and parameters'
       );
     }
     if (forkLevel && !validForkLevels.includes(forkLevel)) {
       return Promise.reject(new GeneralError(`fork level must be one of: ${validForkLevels.join()}`));
     }
-    return test(id, forkLevel, includeUnmodified, verbose);
+    return test(id, forkLevel, all, verbose);
   }
 
   report(results: SpecsResultsWithComponentId): string {